+++ backend/src/main/java/moadong/club/service/ClubClickService.java
@@ -5,10 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import moadong.club.payload.response.ClubClickRankingResponse;
 import moadong.club.payload.response.ClubClickRankingResponse.ClubRankItem;
 import moadong.club.payload.response.ClubClickResponse;
-import net.javacrumbs.shedlock.spring.annotation.SchedulerLock;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
 import java.time.ZoneId;
@@ -23,11 +20,10 @@ import java.util.concurrent.atomic.AtomicInteger;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-@ConditionalOnProperty(name = "scheduling.enabled", havingValue = "true", matchIfMissing = true)
 public class ClubClickService {
 
-    private static final String CLICK_KEY_PREFIX = "club:click:";
-    private static final String CLICK_KEY_PATTERN = CLICK_KEY_PREFIX + "*";
+    static final String CLICK_KEY_PREFIX = "club:click:";
+    static final String CLICK_KEY_PATTERN = CLICK_KEY_PREFIX + "*";
     private static final ZoneId KST = ZoneId.of("Asia/Seoul");
 
     private final StringRedisTemplate stringRedisTemplate;
@@ -62,17 +58,7 @@ public class ClubClickService {
         return new ClubClickRankingResponse(ranked, nextMidnightKst());
     }
 
-    @Scheduled(cron = "0 0 0 * * *", zone = "Asia/Seoul")
-    @SchedulerLock(name = "ClubClickReset", lockAtMostFor = "1m", lockAtLeastFor = "1s")
-    public void resetDailyClicks() {
-        Set<String> keys = stringRedisTemplate.keys(CLICK_KEY_PATTERN);
-        if (keys != null && !keys.isEmpty()) {
-            stringRedisTemplate.delete(keys);
-        }
-        log.info("동아리 클릭 수 초기화 완료");
-    }
-
-    private String nextMidnightKst() {
+    public String nextMidnightKst() {
         ZonedDateTime nextMidnight = ZonedDateTime.now(KST)
                 .toLocalDate()
                 .plusDays(1)
